requirements.txt
pandas
numpy
scikit-learn
joblib
yfinance
pyarrow
python-dotenv
supabase
PyYAML
